ml-training/requirements.txt
# Core ML Framework
ultralytics==8.0.232
torch==2.1.2
torchvision==0.16.2

# Data Processing
numpy==1.24.3
pandas==2.1.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Visualization & Analysis
matplotlib==3.8.2
seaborn==0.13.0
tensorboard==2.15.1

# Model Export & Optimization
onnx==1.15.0
onnx-simplifier==0.4.36
openvino-dev==2023.2.0

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
scipy==1.11.4
scikit-learn==1.3.2

# Development
jupyter==1.0.0
ipython==8.18.1
pytest==7.4.3
